+++ src/components/Sensitive.svelte
@@ -23,10 +23,10 @@ if (sensitive) {
 			<p>{t("sensitive.warning")}</p>
 		</div>
 		<div class="flex gap-3">
-			<button class="font-bold c-background bg-red-5 py-1 px-2 rd-md" onclick={() => (sensitive = false)}>
+			<button class="font-bold text-background bg-red-500 py-1 px-2 rounded-md" onclick={() => (sensitive = false)}>
 				{t("sensitive.continue")}
 			</button>
-			<a href={getRelativeLocaleUrl(locale, back)} class="flex items-center font-bold c-background bg-secondary py-1 px-2 rd-md">
+			<a href={getRelativeLocaleUrl(locale, back)} class="flex items-center font-bold text-background bg-secondary py-1 px-2 rounded-md">
 				{t("sensitive.back")}
 			</a>
 		</div>